Added Type system
Type system allows the creation of custom Types, characterized by hexcode id number and String name value. Cards can have up to two active Types at once. These two active Types are held on two Type stacks. (planned to be generalized with Type stack array to an arbitrary number of Type stacks). The top Type of the stack is the active Type, and is the one utilized in play. Types can have subTypes. During play, Types on a card can be evaluated into subTypes. This pushes the subType onto the stack where the superType was located on the Card. This replaces the Card's active Type. If a Card has permanence, it retains its evaluation after discard pile resets. If a Card does not have permanence, the two Type stacks are popped until the original active Types are restored (planned to be generalized to selective stack restoration).

TypeSpaces allow the establishing of relations among Types inhabiting it. The relation is modeled as A -> B. TypePools are an orderly way of adding and removing Types for later usage, particularly for addition to TypeSpaces. Any given TypePool can have any number of TypeSpaces derived from it, and I plan to allow the use of multiple TypeSpaces in play. The TypeSpace relation A->B in the main context determines whether a Card with Type A can be discarded onto a Card with Type B. However, the framework is general and versatile, greater customization is expected.

Certain preset types will have special graphics, but in general a graphics engine will utilize either a Type's hexcode (for color) or name (simple text) This process is based on which stack a Type is in, will refine when stack number is generalized.

Next push will probably be the Action system. Actions can be embedded in Cards or Turn phases. Embedded Actions are activated by discading the Card or clicking an on-screen button respectively.